requirements.txt
Flask
beautifulsoup4
flask_restful
flask_jsonpify
request
Pillow
